[FIX] Wrong Negation Join Implementation Under Incremental Mode (#24)
* (fix)[compiler]: wrong negation logic

* (fix)[compiler]: seperate logic of neg join for batch and incremental

* (fix)[compiler]: incomplete semiring imports tracking

* (fix)[compiler]: fix incorrect imports under incremental mode

* (fix)[compiler]: wrong order of render imports

+++ crates/compiler/src/import.rs
@@ -279,7 +279,7 @@ impl ImportTracker {
     /// Differential dataflow diff type.
     fn diff_type(&self) -> TokenStream {
         match self.mode {
-            ExecutionMode::Incremental => quote! { type Diff = isize; },
+            ExecutionMode::Incremental => quote! { type Diff = i32; },
             ExecutionMode::Batch => {
                 quote! { type Diff = differential_dataflow::difference::Present; }
             }
